add more tests with hypothesis and scheduled actions
