+++ .github/hypothesis.md
@@ -1,0 +1,7 @@
+---
+title: Failure in extended hypothesis testing
+labels: bug
+---
+
+The extended tests with hypothesis failed on date {{ date | date('dddd, MMMM Do') }}.
+See [workflow logs](https://github.com/arviz-devs/arviz-plots/actions/workflows/hypothesis.yml) for details on which tests failed and why.
