+++ .github/workflows/hypothesis.yml
@@ -1,0 +1,30 @@
+name: Run extended tests with hypothesis
+on:
+  schedule:
+    - cron: '17 5 * * 1'
+  workflow_dispatch:
+
+jobs:
+  hypothesis_testing:
+    runs-on: ubuntu-latest
+  steps:
+    - uses: actions/checkout@v4
+    - name: Set up Python
+      uses: actions/setup-python@v5
+      with:
+        python-version: '3.11'
+    - name: Install arviz-plots
+      run: |
+        python -m pip install ".[test]"
+    - name: Execute tests
+      run: |
+        pytest --hypothesis-profile chron -k hypothesis
+    - name: Open issue if failed
+      if: failure()
+      uses: JasonEtco/create-an-issue@v2
+      env: 
+        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+      with:
+        filename: .github/hypothesis.md
+        update_existing: true
+        search_existing: all
