+++ tests/conftest.py
@@ -4,8 +4,13 @@ import logging
 import os
 
 import pytest
+from hypothesis import settings
 
 _log = logging.getLogger("arviz_plots")
+
+settings.register_profile("fast", deadline=1000, max_examples=20)
+settings.register_profile("chron", deadline=1000, max_examples=500)
+settings.load_profile(os.getenv("HYPOTHESIS_PROFILE", "fast"))
 
 
 def pytest_addoption(parser):
